requirements.txt
pandas==2.0.3
pymysql==1.1.0
boto3==1.28.0
sqlalchemy==2.0.20
pyyaml==6.0
